Remove ownership of I2C and pass as argument to functions instead.
If an instance of HT16K33 owns the I2C instance, then it's not
possible to control multiple HT16K33 instances over the same I2C line.
This is true when communicating with multiple devices even if they're
not HT16K33 objects.

This change allows multiple devices to be communicating over the same
I2C line.